// src/api/models/ClienteInternoResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClienteInternoResponse = {
    id?: number;
    usuarioId?: number;
    username?: string | null;
    nombres?: string | null;
    apellidos?: string | null;
    nombreCompleto?: string | null;
    genero?: string | null;
    fechaNacimiento?: string;
    telefono?: string | null;
    correoElectronico?: string | null;
    rol?: string | null;
    tipoDocumento?: string | null;
    numeroIdentificacion?: string | null;
    fechaCreacion?: string;
    fechaActualizacion?: string | null;
    usuarioActivo?: boolean;
    clienteActivo?: boolean;
};